Reworked template wrapping to not be a middleware
